requirements.txt
Flask==3.0.0
flask_jwt_extended==4.6.0
passlib==1.7.4
PyJWT==2.8.0
PyJWT==2.8.0
WTForms==3.1.2
pycryptodome==3.20.0
email_validator==2.1.0
Flask-Mail==0.9.1
waitress==2.1.2
